# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Counterpoise API",
    "description": "The routes a native client uses. Money is integer cents. Shares and prices are integer micros. Dates are YYYY-MM-DD strings.",
    "version": "1.48.0",
    "x-api-contract": 1
  },
  "servers": [
    {
      "url": "/",
      "description": "The server this document was fetched from."
    }
  ],
  "paths": {
...
    "/api/b/{bookId}/transactions": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Transactions, newest first. With includeMeta=true the response is a TransactionPage.",
        "operationId": "listTransactions",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "accountId",
            "in": "query",
            "description": "Only transactions with a split in this account.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "accountIds",
            "in": "query",
            "description": "Comma-separated account ids.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "balanceAccountId",
            "in": "query",
            "description": "Account for startingBalance. Defaults to the first of accountIds.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "payeeId",
            "in": "query",
            "description": "Only transactions with this payee.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "recurringRuleId",
            "in": "query",
            "description": "Only transactions this rule created.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Inclusive lower bound as YYYY-MM-DD.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Inclusive upper bound as YYYY-MM-DD.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeMeta",
            "in": "query",
            "description": "Send true to receive a TransactionPage instead of a TransactionList.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size. Default 100. 0 means everything.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Rows to skip.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "ensureId",
            "in": "query",
            "description": "Extend the first page until it includes this transaction.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TransactionList"
                    },
                    {
                      "$ref": "#/components/schemas/TransactionPage"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error. The body carries a message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Create a transaction. Splits must sum to zero.",
        "operationId": "createTransaction",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "default": {
            "description": "Error. The body carries a message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountRow": {
        "type": "object",
        "required": [
          "id",
          "bookId",
          "name",
          "type",
          "subtype",
          "parentId",
          "isActive",
          "isFavorite",
          "isInvestmentCash",
          "icon",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "bookId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "subtype": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountSubtype"
              },
              {
                "type": "null"
              }
            ]
          },
          "parentId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isInvestmentCash": {
            "type": "boolean"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "A single emoji, or null to inherit the parent's icon."
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
      "AccountSubtype": {
        "type": "string",
        "enum": [
          "bank",
          "credit_card",
          "loan",
          "investment",
          "cash",
          "other"
        ]
      },
      "AccountType": {
        "type": "string",
        "enum": [
          "asset",
          "liability",
          "equity",
          "income",
          "expense"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A message safe to show to the user."
          }
        }
      },
...
      "Cents": {
        "type": "integer",
        "description": "Money in integer cents. Positive is a debit, negative is a credit."
      },
...
      "CreateTransactionBody": {
        "type": "object",
        "required": [
          "date",
          "splits"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "Transaction date in YYYY-MM-DD format"
          },
          "description": {
            "type": "string",
            "description": "Transaction description"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          },
          "payeeName": {
            "type": "string",
            "description": "Payee name. Matched case-insensitively to an existing payee in this book; created if none matches."
          },
          "checkNumber": {
            "type": "string",
            "description": "Check number (only for bank account transactions)"
          },
          "isFloating": {
            "type": "boolean",
            "description": "If true, the transaction's effective date auto-advances to today until reconciled. Useful for checks or expected deposits where the clearing date is unknown."
          },
          "isReconciled": {
            "type": "boolean",
            "description": "If true, the transaction is recorded as already reconciled. If the transaction is floating (isFloating is true), also set isFloating to false and give the cleared date in the same call. If you set only isReconciled, the transaction stays floating, and its effective date keeps moving to today even though it is marked reconciled."
          },
          "splits": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "accountId",
                "amount"
              ],
              "properties": {
                "accountId": {
                  "type": "integer",
                  "description": "The account ID for this split",
                  "exclusiveMinimum": 0
                },
                "amount": {
                  "type": "integer",
                  "description": "Amount in cents. Positive = debit, negative = credit. All splits must sum to zero."
                }
              }
            },
            "description": "Transaction splits (must sum to zero)",
            "minItems": 2
          },
          "investmentSplits": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "securityId",
                "action",
                "sharesMicros",
                "priceMicros"
              ],
              "properties": {
                "securityId": {
                  "type": "integer",
                  "description": "The security ID",
                  "exclusiveMinimum": 0
                },
                "action": {
                  "type": "string",
                  "description": "Investment action type",
                  "enum": [
                    "buy",
                    "sell",
                    "dividend",
                    "capGain",
                    "fee",
                    "split"
                  ]
                },
                "sharesMicros": {
                  "type": "integer",
                  "description": "Number of shares in micros (1,000,000 = 1 share). Always positive.",
                  "minimum": 0
                },
                "priceMicros": {
                  "type": "integer",
                  "description": "Price per share in micros (1,000,000 = $1.00)",
                  "minimum": 0
                },
                "feesCents": {
                  "type": "integer",
                  "description": "Fees in cents",
                  "minimum": 0
                },
                "splitNumerator": {
                  "type": "integer",
                  "description": "For stock split: new share ratio numerator",
                  "exclusiveMinimum": 0
                },
                "splitDenominator": {
                  "type": "integer",
                  "description": "For stock split: old share ratio denominator",
                  "exclusiveMinimum": 0
                }
              }
            },
            "description": "Investment splits (required if transaction involves investment accounts)"
          }
        }
      },
      "InvestmentSplitRow": {
        "type": "object",
        "required": [
          "id",
          "bookId",
          "transactionId",
          "accountId",
          "securityId",
          "action",
          "sharesMicros",
          "priceMicros",
          "feesCents",
          "splitNumerator",
          "splitDenominator",
          "security",
          "account"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "bookId": {
            "type": "integer"
          },
          "transactionId": {
            "type": "integer"
          },
          "accountId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "securityId": {
            "type": "integer"
          },
          "action": {
            "type": "string",
            "enum": [
              "buy",
              "sell",
              "dividend",
              "capGain",
              "fee",
              "split"
            ]
          },
          "sharesMicros": {
            "$ref": "#/components/schemas/Micros"
          },
          "priceMicros": {
            "$ref": "#/components/schemas/Micros"
          },
          "feesCents": {
            "$ref": "#/components/schemas/Cents"
          },
          "splitNumerator": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "splitDenominator": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "security": {
            "$ref": "#/components/schemas/SecurityRow"
          },
          "account": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountRow"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "LedgerDate": {
        "type": "string",
        "format": "date",
        "description": "A calendar date as YYYY-MM-DD. No time zone."
      },
...
      "Micros": {
        "type": "integer",
        "description": "Shares or a price in micros. 1,000,000 micros is one share or one dollar."
      },
...
      "PayeeRow": {
        "type": "object",
        "required": [
          "id",
          "bookId",
          "name",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "bookId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
...
      "SecurityRow": {
        "type": "object",
        "required": [
          "id",
          "bookId",
          "name",
          "symbol",
          "securityType",
          "fetchPrices",
          "fixedPriceMicros",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "bookId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "securityType": {
            "type": "string",
            "enum": [
              "etf",
              "mutual_fund",
              "stock"
            ]
          },
          "fetchPrices": {
            "type": "boolean"
          },
          "fixedPriceMicros": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Micros"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
...
      "Timestamp": {
        "type": "string",
        "format": "date-time",
        "description": "A UTC instant as an ISO 8601 string with a Z suffix."
      },
      "Transaction": {
        "type": "object",
        "required": [
          "id",
          "bookId",
          "date",
          "description",
          "checkNumber",
          "notes",
          "payeeId",
          "isReconciled",
          "isFloating",
          "recurringRuleId",
          "createdAt",
          "updatedAt",
          "payee",
          "splits",
          "investmentSplits"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "bookId": {
            "type": "integer"
          },
          "date": {
            "$ref": "#/components/schemas/LedgerDate",
            "description": "The stored date. For a floating transaction the effective date is today; compute it on the client."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "checkNumber": {
            "type": [
              "string",
              "null"
            ]
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "payeeId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "isReconciled": {
            "type": "boolean"
          },
          "isFloating": {
            "type": "boolean"
          },
          "recurringRuleId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdBy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The user who created this row. An older client contract omits this field. Null means a system write."
          },
          "updatedBy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The user who last changed this row. An older client contract omits this field. Null means a system write."
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "payee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PayeeRow"
              },
              {
                "type": "null"
              }
            ]
          },
          "splits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionSplit"
            }
          },
          "investmentSplits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvestmentSplitRow"
            }
          }
        }
      },
      "TransactionList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Transaction"
        }
      },
      "TransactionPage": {
        "type": "object",
        "required": [
          "transactions",
          "startingBalance",
          "totalCount"
        ],
        "properties": {
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          },
          "startingBalance": {
            "$ref": "#/components/schemas/Cents",
            "description": "Sum of the balance account's splits older than the oldest transaction on this page."
          },
          "totalCount": {
            "type": "integer"
          }
        }
      },
      "TransactionSplit": {
        "type": "object",
        "required": [
          "id",
          "bookId",
          "transactionId",
          "accountId",
          "amount",
          "account"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "bookId": {
            "type": "integer"
          },
          "transactionId": {
            "type": "integer"
          },
          "accountId": {
            "type": "integer"
          },
          "amount": {
            "$ref": "#/components/schemas/Cents"
          },
          "account": {
            "$ref": "#/components/schemas/AccountRow"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key: Authorization: Bearer cpk_..."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "counterpoise_session"
      }
    }
  }
}